[ipb] Ingame Performance Browser. Enable with /perfbrowse.
Works but the columns get messed up if you change category due to a bug in guiGridListRemoveColumn which I'll be reporting now and this resource can be used as it contains a script to test the bug.

How it should look: http://cit2.net/arran/dev/ipb1.jpg

This is just basic for now. Options, filter and target boxes aren't done. The idea isn't just an ingame copy of performancebrowser but to add an option to save the stats so that if something goes wrong and you're not staring at performancebrowser you'll be able to see what happened.

Also being able to view graphs of important things like player count and server logic FPS are planned, but for now it's got this grid list bug.

git-svn-id: https://mtasa-resources.googlecode.com/svn/trunk@896 126a8462-9353-1474-c972-ef1b1113b2d6